modify

+++ Probleme_solving/sem1/gestion_livres/action.php
@@ -1,17 +1,25 @@
 <?php
 include_once('conn.php');
-$code=$_POST["code"];
-$lib=$_POST["lib"];
-$aut=$_POST["aut"];
-$maison=$_POST["maison"];
-$nbt=$_POST["nbt"];
-$res=false;
+if(isset($_POST["code"])){
+    $code=$_POST["code"];
+    $lib=$_POST["lib"];
+    $aut=$_POST["aut"];
+    $maison=$_POST["maison"];
+    $nbt=$_POST["nbt"];
+    if(isset($_POST["res"])){
+        $res=$_POST["res"] ? 1 : 0;
+    }
+    else{
+    $res=0;
+    }
+}
 if(isset($_POST["Ajouter"])){
     $sel=$conn->prepare("SELECT * FROM livres where code=?");
     $sel->bindValue(1,$code);
     $sel->execute();
     if($sel->rowCount()>0){
-        echo"livre deja existe";
+         ?><script>alert("<?php echo"livre deja existe" ;?>")</script>
+        <?php
     }
     else{
         $query=$conn->prepare("INSERT INTO livres values(?,?,?,?,?,?)");
@@ -31,13 +39,14 @@ if(isset($_POST["Ajouter"])){
     }
 }
 else if(isset($_POST["Modifier"])){
-    $query=$conn->prepare("UPDATE livres set code=?,libelle=?,auteur=?,maison_edit=?,nb_tomes=? where code=?");
+    $query=$conn->prepare("UPDATE livres set code=?,libelle=?,auteur=?,maison_edit=?,nb_tomes=?,reserver=? where code=?");
     $query->bindValue(1,$code);
     $query->bindValue(2,$lib);
     $query->bindValue(3,$aut);
     $query->bindValue(4,$maison);
     $query->bindValue(5,$nbt);
-    $query->bindValue(6,$code);
+    $query->bindValue(6,$res);
+    $query->bindValue(7,$code);
     if($query->execute()){
             header("Location: index.php");
             exit();
@@ -47,7 +56,7 @@ else if(isset($_POST["Modifier"])){
         }
 
 }
-else{
+else if(isset($_POST["Supprimer"])){
     $query=$conn->prepare("DELETE FROm livres where code=?");
     $query->bindValue(1,$code);
     if($query->execute()){
@@ -55,10 +64,26 @@ else{
         exit();
         }
     else{
-            echo "error";
+            ?><script>alert("<?php echo"erreur" ;?>")</script>
+        <?php
         }
 }
-
+else if(isset($_POST["recherche"])){
+    if(isset($_POST["code"])){
+    $q=$conn->prepare("SELECT * from livres where code=?");
+    $q->bindValue(1,$_POST["code"]);
+    $q->execute();
+    if($q->rowCount()>0){
+        $rq=$q->fetch();
+    }
+    else{
+        ?><script>alert("<?php echo"le livre ne trouve pas" ;?>")</script>
+        <?php
+        ;
+    }
+    
+}
+}
